fixed current location on mapbox
