+++ api_dev/dropin-frontend/src/pages/map-view/index.jsx
@@ -1,95 +1,15 @@
-// import React, { useRef, useEffect, useState } from 'react';
-// import LocationComponent from "../../components/Location/Location";
-// import useLocation from "../../custom-hooks/useLocation";
-// import mapboxgl from '!mapbox-gl'; // eslint-disable-line import/no-webpack-loader-syntax
-//
-// mapboxgl.accessToken = 'pk.eyJ1Ijoiam9zZWZvMTk5NyIsImEiOiJjbGp5emxjYncwMWxhM2dvNDdmaWp5dWFkIn0.NlylS3S5YnEYHQ6JyAR-HA';
-//
-// function MapView() {
-//     return (
-//         <>
-//             <p>Map View Page</p>
-//             <LocationComponent/>
-//             <SimpleMap/>
-//         </>
-//     );
-// }
-//
-// function SimpleMap() {
-//     const mapContainer = useRef(null);
-//     const map = useRef(null);
-//     const [lng, setLng] = useState(-70.9);
-//     const [lat, setLat] = useState(42.35);
-//     const [zoom, setZoom] = useState(9);
-//
-//     useEffect(() => {
-//         if (map.current) return; // initialize map only once
-//         map.current = new mapboxgl.Map({
-//             container: mapContainer.current,
-//             style: 'mapbox://styles/mapbox/streets-v12',
-//             center: [lng, lat],
-//             zoom: zoom
-//         });
-//     })
-//
-//     useEffect(() => {
-//         if (!map.current) return; // wait for map to initialize
-//         map.current.on('move', () => {
-//             setLng(map.current.getCenter().lng.toFixed(4));
-//             setLat(map.current.getCenter().lat.toFixed(4));
-//             setZoom(map.current.getZoom().toFixed(2));
-//         });
-//     });
-//
-//     return (
-//         <div>
-//             <div className="sidebar">
-//                 Longitude: {lng} | Latitude: {lat} | Zoom: {zoom}
-//             </div>
-//             <div ref={mapContainer} className="map-container"/>
-//         </div>
-//     );
-// }
-//
-//
-// export default SimpleMap;
-import React, { useRef, useEffect, useState } from 'react';
+import React, {useRef, useEffect, useState} from 'react';
 import mapboxgl from '!mapbox-gl'; // eslint-disable-line import/no-webpack-loader-syntax
+import useLocation from "../../custom-hooks/useLocation";
+import SimpleMap from "../../components/Map/SimpleMap";
 
-mapboxgl.accessToken = 'pk.eyJ1Ijoiam9zZWZvMTk5NyIsImEiOiJjbGp5emxjYncwMWxhM2dvNDdmaWp5dWFkIn0.NlylS3S5YnEYHQ6JyAR-HA';
-
-export default function App() {
-  const mapContainer = useRef(null);
-  const map = useRef(null);
-  const [lng, setLng] = useState(70.9);
-  const [lat, setLat] = useState(42.35);
-  const [zoom, setZoom] = useState(9);
-
-  useEffect(() => {
-    if (map.current) return; // initialize map only once
-    map.current = new mapboxgl.Map({
-      container: mapContainer.current,
-      style: 'mapbox://styles/mapbox/streets-v12',
-      center: [lng, lat],
-      zoom: zoom
-    });
-  })
-
-  useEffect(() => {
-    if (!map.current) return; // wait for map to initialize
-    map.current.on('move', () => {
-      setLng(map.current.getCenter().lng.toFixed(4));
-      setLat(map.current.getCenter().lat.toFixed(4));
-      setZoom(map.current.getZoom().toFixed(2));
-    });
-  });
-
-  return (
-    <div align="center">
-      <div className="sidebar">
-        Longitude: {lng} | Latitude: {lat} | Zoom: {zoom}
-      </div>
-      <div ref={mapContainer} className="map-container"  />
-    </div>
-  );
+function MapView() {
+    const userLocation = useLocation();
+    return (
+        <>
+            <SimpleMap userLocation={userLocation}/>
+        </>
+    );
 }
+
+export default MapView;
